moved go.mod to /v2 path on the main.
