+++ mongodb/test/data/rds-combined-ca-bundle.pem
