--- v2/go.mod
@@ -1,19 +1,0 @@
-module github.com/ONSdigital/dp-mongodb/v2
-
-go 1.16
-
-require (
-	github.com/ONSdigital/dp-component-test v0.3.1
-	github.com/ONSdigital/dp-healthcheck v1.0.5
-	github.com/ONSdigital/dp-mongodb v1.5.0
-	github.com/ONSdigital/dp-net v1.0.9 // indirect
-	github.com/ONSdigital/log.go v1.0.1
-	github.com/cucumber/godog v0.10.0
-	github.com/globalsign/mgo v0.0.0-20181015135952-eeefdecb41b8 // indirect
-	github.com/go-test/deep v1.0.7 // indirect
-	github.com/smartystreets/goconvey v1.6.4
-	github.com/square/mongo-lock v0.0.0-20201208161834-4db518ed7fb2
-	github.com/stretchr/testify v1.6.1
-	go.mongodb.org/mongo-driver v1.5.2
-	gopkg.in/check.v1 v1.0.0-20200902074654-038fdea0a05b // indirect
-)
